+++ src/controllers/resources/form/products.js
@@ -2135,5 +2135,8 @@ export default function () {
     setTimeout(() => {
       listOrders(tabId)
     }, 1500)
+    document.getElementById(`t${tabId}-ad-relevance-link`).addEventListener('click', () => {
+      $(`#t${tabId}-ad-relevance`).slideToggle()
+    }, false)
   }
 }