+++ src/router/load-route.js
@@ -21,6 +21,8 @@ export default path => {
       return import(/* webpackChunkName: "router_routes_inventory" */ '@/router/routes/cashback')
     case 'campaign-report':
       return import(/* webpackChunkName: "router_routes_campaign-report" */ '@/router/routes/campaign-report')
+    case 'coupon-report':
+      return import(/* webpackChunkName: "router_routes_coupon-report" */ '@/router/routes/coupon-report')
     case 'device-report':
       return import(/* webpackChunkName: "router_routes_device-report" */ '@/router/routes/device-report')
     case 'freight-report':